Add search to User * Move search logics to PaintingSearch component
